:recycle: Moved preventTooLongInput in validation utility
