+++ sitzverteilung-frontend/src/components/basedata/groupdata/GroupDataTableRow.vue
@@ -75,8 +75,10 @@ import type { VTextField } from "vuetify/components";
 
 import { computed, nextTick, useTemplateRef, watch } from "vue";
 
-import { preventTooLongInput } from "@/utility/input.ts";
-import { FieldValidationRules } from "@/utility/validation.ts";
+import {
+  FieldValidationRules,
+  preventTooLongInput,
+} from "@/utility/validation.ts";
 
 const {
   isValidatingOnEmpty = true,
